Create MP Config TCK Runner
This change creates a MP Config TCK runner module to run the TCK
against a managed instance of Glassfish. There are existing dependency
issues, which means there are a decent proportion of tests that don't
run correctly. At this point it's fine to test that the basic deployment
of MP Config apps works as expected.

Each MP TCK module will have its own instance of Glassfish, but will be
passed a common arquillian.xml.

The TCK requires a Weld impl to run due to the Arquillian CDI enricher.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>