Replace beans.xml files in MP Config TCK
The TCK wrongly assumes that, like in CDI versions <4, the 'all'
bean discovery mode is assumed. GlassFish is now CDI 4.0 compliant,
which means that the 'all' bean discovery mode now needs to be
manually specified. To fix this, the arquillian extension now replaces
the beans.xml files with custom versions. This will need modifying if
the TCK contains any meaningful beans.xml tests.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>