Cast GlassfishClientExceptions to expected exception type
When using managed and remote clients, the Arquillian GF connectors throw
each deployment exception as a GlassfishClientException, with the body of
the actual exception. For the TCK to expect a certain exception type, we
need an extension to interpret the response as the correct type. This
change adds an extension to convert the "CDI deployment failure"
responses to jakarta namespaced DeploymentExceptions, as the server produces.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>