Configure the required container system properties
System properties should ideally be configured via the arquillian
container. This isn't possible via the official managed extension,
so a workaround is needed. The options were either to configure it
via a remote Arquillian extension (which doesn't currently work as
Smallrye initialises its config sources before these extensions
run), or to run a new test which sets it up. This change creates a
new test purely to setup the container programmatically.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>