Modify non-testable TCK deployments
The MP TCK uses ShouldThrowException to check some deployment failures.
By default, Arquillian will not apply the extension services to those
deployments. The beans.xml processor is still required in some cases
though, to force the container to create bean deployments for the app.
This change applies the processer during the BeforeDeploy event when
the archive is non testable.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>