Allow sniffers to scan annotated parameters
MicroProfile annotations are allowed to be injected into method parameters,
but the sniffer code currently tries to cast the parameter to a member
(i.e. field or method), which causes a ClassCastException. This change
modifies the sniffer code to allow the sniffer to find the class containing
the parameter annotation.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>