Eagerly create MDBs on startup
The CDI TCK requires events for MDBs to be fired on startup. The easiest
way to satisfy that requirement is to eagerly load MDBs by default.
Users can still disable that if needed.

First commit is some refactoring, mostly formatting and some naming.
Second commmit will be the actual change.

Signed-off-by: Arjan Tijms <arjan.tijms@gmail.com>