Prevent Helidon microprofile-config.properties caching
Smallrye used a custom file loading mechanism that doesn't have this bug.
Since Helidon config uses JarURLConnection to load the microprofile-config.properties
file, applications with the same name will return the same cached file contents.

Disabling it server-wide would be a big performance hit, so instead this fix disables
the cache during the config provider initialisation, and then returns it to the default
behaviour. This shouldn't cause performance problems as I expect the file to only be
read from the same archive once.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>