Copy the ConfigCdiExtension into the config module
The embedded dependencies needed not inlining, as by default
they overwrite classes from the main module. This isn't the case
when using the 'unpackBundle' option, but that causes module-info
files from the dependencies to break compilation. A sensible option
seems to be to not inline the dependencies.

Note that the overwritten class causes the single artifact it comes
from to be inlined.

Signed-off-by: Matthew Gill <matthew.gill@live.co.uk>