GH #25039 Facelets in Console: Moved integration to the commandrecorder plugin
Moved all the code needed to integrate the Command Recorder console plugin into the plugin module. This means that Admin Console will work without issues if the module is not present.

Since the Admin Recorder plugin is deployed in the AdminGUI application and not as an OSGi bundle in modules, I needed to update the ConsolePluginService:

* it's now created by an HK2 locator specific to the deployed app, so that it finds console plugins also in the web app, not only in OSGi modules
* it searches for all plugin configuration files in the classloader and remove duplicities (because the web app classloader also finds all files in OSGi bundles via the parent classloader)

Also removed the demo facelets files.(header.xhtml, etc.)