HK2 JUnit5 extension: Start HK2 topic service, clean up global habitat
The first time the global service locator is accessed, the current locator is registered as the global one. We need to reset it so that further tests start with a clean environment. We do this in the HK2JUnit5Extension, moved to the hk2 subpackage, while all the rest is moved to the base-test-utils module.

I needed to split the HK2 extension into 2 modules, because the dependency on internal GF API module introduced a cyclic dependency. The GF API module (via the config-api dependency) now depends only on the base-test-module to avoid the cyclic dependency.