Reload context classloader for each config change.
Because common CL is lazily initialized, it can point to its parent for a while. Thread context CL for config change handlers should be refreshed before each task, in case common CL was updated after the executor was created.

This is especially useful in GF Embedded if add-library is called at startup. It used to happen that the config change executor was created before the common CL was updated by the add-library command.