Fixes in runnable GlassFish Embedded for add-library command (#25420)
* Do not log exception on add-library called in GF Embedded

If domain.xml file modifications disabled, don't try to touch the file.

* Reload context classloader for each config change.

Because common CL is lazily initialized, it can point to its parent for a while. Thread context CL for config change handlers should be refreshed before each task, in case common CL was updated after the executor was created.

This is especially useful in GF Embedded if add-library is called at startup. It used to happen that the config change executor was created before the common CL was updated by the add-library command.

* Refactor config test to use non-deprecated methods

* Add tests for runnable GF Embedded

For add-library - when a JAR is used by an app, and also by GF itself (to configure Grizzly)