Fix cleanup on deployment error
Before this, JNDI bindings were not cleaned up for all EJBs, leaking information to future deployments of the same name.
If an app didn't deploy successfully and was deployed later, JNDI bindings remained from the previous failed deployment instead of being initialized from scratch. I saw an issue that EntityManagerFactory was closed because of a failed deployment, and then it was reused for the next deployment, leading to a closed EMF assigned to a newly deployed app. 

The cause for this was that the JNDI bindings were removed only after they were unregistered the same number of times as they were registered. During failed deployment, they were not unregistered in some cases, and an old JNDI reference was kept in the map. The fix removes all the bindings once again during application unregistration.