Jakarta Data integration - Upgrade Data to 1.1.10-SNAPSHOT
In JNoSQL 1.1.9/1.1.10-SNAPSHOT, ClassScanner is cached as a singleton. We need to retrieve fresh types.

Always update Maven Central Snapshots only in the "ci" profile. 
In CI, this is needed to attempt redownload snaphots immediately after the snapshot dependency is released. 
In dev it's not needed, it only slows down local build.