Refactor Embedded Web to use FeatureSets
We don't use GlassFish build-helper plugin, we need to exclude some dependencies in the feature sets we don't want. Build-helper plugin doesn't need it, it only takes first level dependencies and ignores transitiva dependencies. But it unpacks JARs into separate directories, which isn't useful, we want then shaded in a single directory.

glassfish-jul dependency must be set as compile. It's set as test scope by default in the glassfish root parent, and it doesn't help to define it as a compile dep in a featureset.