Jakarta Data Integration - Update NoSQL reflection mechanism
Synchronize with code in 1.1.11.
Injecting by class, not by interface as in the original JNoSQL code. This ensures that our implementation is always injected and a bean from an application library is never preferred. This is necessary because enabling alternative is not enough - the JNoSQL reflection module, which is a transitive maven dependency of JNoSQL drivers, also defines its bean as enabled alternative and takes precendence. This then causes classloading failures, as the GlassFish classloader doesn't see the classes from the app.

I also removed declaration of class scanner as CDI bean, it's not injected anywhere, it's loaded by service loader.

Current status:

* Changes from ondromih-8.0-mp-config-per-app branch need to be merged
* These tests are failing for Data TCK with Mongo:

Failures: 
  EntityTests.testUpdateQueryWithWhereClause:2489 expected: <2> but was: <0>
  EntityTests.testUpdateQueryWithoutWhereClause:2413 expected: <3> but was: <0>
  ValidationTests.testInvalidResultsFromMethod:311 Rectangles.findEverythingIdSorted did not raise ConstraintViolationException.
  ValidationTests.testSaveWithInvalidConstraints:102 Expected jakarta.validation.ConstraintViolationException to be thrown, but nothing was thrown.
  ValidationTests.testUpdateAllWithInvalidConstraints:240 Expected jakarta.validation.ConstraintViolationException to be thrown, but nothing was thrown.
  ValidationTests.testUpdateWithInvalidConstraints:202 Expected jakarta.validation.ConstraintViolationException to be thrown, but nothing was thrown.

The tests in EntityTests are failing because JNoSQL doesn't return the number of deleted entities from delete queries. Change needed in JNoSQL to support this for databases that support it, or in the TCK do disable these assertions for DBs that don't support it.

ValidationTests - need to be implemented by an interceptor, similarly to JPA impl